+++ broker/bam/inc/com/centreon/broker/bam/ba_ratio_number.hh
@@ -46,8 +46,8 @@ class ba_ratio_number : public ba {
 
  public:
   ba_ratio_number(uint32_t id,
-                  uint32_t host_id,
-                  uint32_t service_id,
+                  uint64_t host_id,
+                  uint64_t service_id,
                   bool generate_virtual_status,
                   const std::shared_ptr<spdlog::logger>& logger);
   state get_state_hard() const override;
