+++ broker/bam/inc/com/centreon/broker/bam/hst_svc_mapping.hh
@@ -37,10 +37,10 @@ using com::centreon::common::log_v2::log_v2;
 class hst_svc_mapping {
   std::shared_ptr<spdlog::logger> _logger;
   absl::flat_hash_map<std::pair<std::string, std::string>,
-                      std::pair<uint32_t, uint32_t>>
+                      std::pair<uint64_t, uint64_t>>
       _mapping;
 
-  absl::flat_hash_map<std::pair<uint32_t, uint32_t>, bool> _activated_mapping;
+  absl::flat_hash_map<std::pair<uint64_t, uint64_t>, bool> _activated_mapping;
 
  public:
   hst_svc_mapping(const std::shared_ptr<spdlog::logger>& logger)
@@ -48,18 +48,18 @@ class hst_svc_mapping {
   ~hst_svc_mapping() noexcept = default;
   hst_svc_mapping(const hst_svc_mapping&) = delete;
   hst_svc_mapping& operator=(const hst_svc_mapping&) = delete;
-  uint32_t get_host_id(std::string const& hst) const;
-  std::pair<uint32_t, uint32_t> get_service_id(std::string const& hst,
+  uint64_t get_host_id(std::string const& hst) const;
+  std::pair<uint64_t, uint64_t> get_service_id(std::string const& hst,
                                                std::string const& svc) const;
-  void set_host(std::string const& hst, uint32_t host_id);
+  void set_host(std::string const& hst, uint64_t host_id);
   void set_service(std::string const& hst,
                    std::string const& svc,
-                   uint32_t host_id,
-                   uint32_t service_id,
+                   uint64_t host_id,
+                   uint64_t service_id,
                    bool activated);
-  void set_activated(uint32_t host_id, uint32_t service_id, bool activated);
+  void set_activated(uint64_t host_id, uint64_t service_id, bool activated);
 
-  bool get_activated(uint32_t hst_id, uint32_t service_id) const;
+  bool get_activated(uint64_t hst_id, uint64_t service_id) const;
 };
 }  // namespace com::centreon::broker::bam
